requirements.txt
bitsandbytes
accelerate
peft
transformers
evaluate
pyarrow
torch
ftfy
torchinfo
scikit-learn
pandas
numpy
pyyaml
flet
uvicorn[standard]
python-multipart
beautifulsoup4
deepl
deep_translator